Start moving code of a specific “page” to a separate place in the backend folder
