+++ src/templates.ts
@@ -5,6 +5,7 @@ import { compile, TemplateFunction, Options } from 'ejs';
 
 const headerHtml = readFileSync(robustPath('views/header.html')).toString();
 
+// TODO: remove this function
 export function processTemplate(filename: string, variables: any = {}): string {
   const fileContent = readFileSync(robustPath(`views/${filename}`)).toString();
